Highlight processed reassigned BZ issues (1/2)
If someone from the security team has already worked on a BZ bug which
was previously re-assigned to the security team, then I want to see
that in the same "reassigned back" folder.

This commit only addresses re-reassigned issues, with a future commit
I'll also address if an issue will be closed by the security team.