Highlight processed reassigned BZ issues (2/2)
As an agreement security related issues should not be closed by the
assignee after he solves the issue, instead the issue has to be assigned
back to the security team, who will then review it and close the issue
if everything is ok. This commit is intended to put the closing
notification right after the re-assigned to security-team notification.
This will help me to quickly see which are the BZ issues that are
reassigned back but not closed (reviewed by the security team).